rainbow: v2 first commit
- prg banking OK
- chr banking OK
- NT mapping OK
- prg self-flashing in mode 3 OK
- chr self-flashing in mode 3 OK
- scanline IRQ OK
- CPU Cycle IRQ ok
- multiplication OK
- version OK

+++ src/ppu.h
@@ -16,7 +16,7 @@ void newppu_hacky_emergency_reset();
 /* For cart.c and banksw.h, mostly */
 extern uint8 NTARAM[0x800], *vnapage[4];
 extern uint8 PPUNTARAM;
-extern uint8 PPUCHRRAM;
+extern uint16 PPUCHRRAM;
 
 void FCEUPPU_SaveState(void);
 void FCEUPPU_LoadState(int version);
